--- Cupcake/src/main/java/dk/cupcake/Category.java
@@ -1,36 +1,0 @@
-// Package
-package dk.cupcake;
-
-// Imports
-
-public class Category {
-
-    // Attributes
-    private int id;
-    private String name;
-
-    // ___________________________________________________
-
-    public String getName() {
-        return name;
-    }
-
-    // ___________________________________________________
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    // ___________________________________________________
-
-    public int getId() {
-        return id;
-    }
-
-    // ___________________________________________________
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-} // Category end
