+++ Cupcake/src/main/java/dk/cupcake/entites/CupcakeTopping.java
@@ -1,0 +1,49 @@
+// Package
+package dk.cupcake.entites;
+
+// Imports
+
+public class CupcakeTopping {
+
+    // Attributes
+    private int id;
+    private String name;
+    private double price;
+
+    // ___________________________________________________
+
+    public double getPrice() {
+        return price;
+    }
+
+    // ___________________________________________________
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    // ___________________________________________________
+
+    public int getId() {
+        return id;
+    }
+
+    // ___________________________________________________
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    // ___________________________________________________
+
+    public String getName() {
+        return name;
+    }
+
+    // ___________________________________________________
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+} // CupcakeTopping end
