+++ Cupcake/src/main/java/dk/cupcake/entites/OrderItem.java
@@ -1,5 +1,5 @@
 // Package
-package dk.cupcake;
+package dk.cupcake.entites;
 
 // Imports
