+++ Cupcake/src/main/java/dk/cupcake/mapper/OrderItemMapper.java
@@ -2,7 +2,7 @@
 package dk.cupcake.mapper;
 
 // Imports
-import dk.cupcake.OrderItem;
+import dk.cupcake.entites.OrderItem;
 import dk.cupcake.db.Database;
 import java.sql.*;
 import java.util.ArrayList;
