+++ Cupcake/src/main/resources/static/Postgre/setup.sql
@@ -13,7 +13,7 @@ id SERIAL PRIMARY KEY,
 email VARCHAR(255) UNIQUE NOT NULL,
 password_hash TEXT NOT NULL, /* Safety */
 role VARCHAR(20) NOT NULL DEFAULT 'customer', /* Customer, Admin, Support (?), Diddy */
-username VARCHAR(100) NOT NULL,
+username VARCHAR(100) UNIQUE NOT NULL,
 created_at TIMESTAMP DEFAULT NOW() /* Auto */
 );
