+++ Cupcake/src/main/resources/templates/events.html
@@ -1,0 +1,137 @@
+<!DOCTYPE html>
+<html lang="da">
+<head>
+
+    <!-- Meta -->
+    <meta charset="UTF-8"> <!-- ÆØÅ -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"> <!-- In case some nerd uses Internet Explorer -->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- Mobile Responsiveness -->
+    <meta name="description" content="Fuck vi hader livet jo... #swag"> <!-- SEO / Google -->
+    <meta name="author" content="Gruppe D"> <!-- Author of Website -->
+
+    <!-- Visuals -->
+    <title>Olsker Cupcakes</title>
+
+    <!-- Icon -->
+    <link rel="icon" href="/images/icons/cupcake-1.png" type="image/x-icon"> <!-- That little logo thats next to the title in URL -->
+    <link rel="apple-touch-icon" href="/images/icons/cupcake-1.png">
+
+    <!-- CSS Load -->
+    <link rel="stylesheet" href="/css/style.css">
+
+    <!-- Scripts can also be loaded here with defer -->
+    <!-- Framwork load -->
+    <link rel="stylesheet" href="/css/guacamoleboy-wow.css">
+    <link rel="stylesheet" href="/css/guacamoleboy-framework.css">
+    <script src="/js/guacamoleboy-wow.js" defer></script>
+
+    <!-- Modules -->
+    <script type="module" src="/js/components/navbar.js" defer></script>
+    <script type="module" src="/js/components/footer.js" defer></script>
+    <script type="module" src="/js/components/footer-nav.js" defer></script>
+
+</head>
+<body class="subpage event-page">
+
+    <!-- Navbar -->
+    <div id="navbar-component"></div>
+    <!-- Navbar -->
+
+    <!-- Hero Section Start -->
+    <section class="hero">
+
+        <div class="guac-container">
+            <div class="guac-row">
+                
+                <div class="guac-col-auto subpage-cupcake-1">
+                    <img src="/images/page-content/index/cupcake-3.png" alt="Billede 1">
+                </div>
+                <div class="guac-col-auto subpage-cupcake-2">
+                    <img src="/images/page-content/index/cupcake-3.png" alt="Billede 2">
+                </div>
+
+            </div>
+        </div>
+
+        <!-- End Wave Hero Start -->
+        <div class="wave-1">
+            <img src="/images/waves/wave-top.svg">
+        </div>
+        <!-- End Wave Hero End -->
+
+    </section>
+    <!-- Hero Section End -->
+
+    <!-- Event Section Start -->
+    <section class="section-1">
+        <div class="guac-container">
+            <div class="guac-row">
+
+                <!-- Event Card Start -->
+                <div class="guac-col-12 guac-mt-5">
+                    <h2>Upcoming Events</h2>
+                    <p>Se alle vores kommende events og tilmeld dig i dag!</p>
+                    <div class="guac-grid-3 guac-gap-4">
+
+                        <div class="guac-card event-card">
+                            <div class="event-image-container">
+                                <img src="/images/page-content/events/event-card-1.png" alt="Autumn Cupcake Fair">
+                                <div class="event-badge">12/10-2025</div>
+                            </div>
+                                <div class="guac-card-title">Cupcake Eating Contest</div>
+                                <div class="guac-card-text">
+                                    Whos the fattest of the fattest? 90min. Eat as many as u can, fatty.
+                                </div>
+                            <a href="#" class="guac-btn guac-btn-event guac-mt-2">Join Now</a>
+                        </div>
+
+                        <div class="guac-card event-card">
+                            <div class="event-image-container">
+                                <img src="/images/page-content/events/event-card-1.png" alt="Autumn Cupcake Fair">
+                                <div class="event-badge">20/10-2025</div>
+                            </div>
+                            <div class="guac-card-title">Chocolate Dipping</div>
+                            <div class="guac-card-text">
+                                Dip your balls in chocolate and make a mold for a loved one!
+                            </div>
+                            <a href="#" class="guac-btn guac-btn-event guac-mt-2">Reserve Spot</a>
+                        </div>
+
+                        <div class="guac-card event-card">
+                            <div class="event-image-container">
+                                <img src="/images/page-content/events/event-card-1.png" alt="Autumn Cupcake Fair">
+                                <div class="event-badge">31/10-2025</div>
+                            </div>
+                            <div class="guac-card-title">#BLM Meeting</div>
+                            <div class="guac-card-text">
+                                let's get it! yesssirrrrski. Whatever. This is a text.
+                            </div>
+                            <a href="#" class="guac-btn guac-btn-event guac-mt-2">Learn More</a>
+                        </div>
+
+                    </div>
+                </div>
+                <!-- Event Card End -->
+
+            </div>
+        </div>
+
+        <!-- End Wave Section 2 Start -->
+        <div class="wave-3">
+            <img src="/images/waves/footer-wave.svg">
+        </div>
+        <!-- End Wave Section 2 End -->
+
+    </section>
+    <!-- Event Section End -->
+
+    <!-- Footer -->
+    <div id="footer-component"></div>
+    <!-- Footer -->
+
+    <!-- Footer Nav -->
+    <div id="footer-nav-component"></div>
+    <!-- Footer Nav -->
+
+</body>
+</html>
