+++ Cupcake/src/main/resources/templates/jobs.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html lang="da">
+<head>
+
+    <!-- Meta -->
+    <meta charset="UTF-8"> <!-- ÆØÅ -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"> <!-- In case some nerd uses Internet Explorer -->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- Mobile Responsiveness -->
+    <meta name="description" content="Fuck vi hader livet jo... #swag"> <!-- SEO / Google -->
+    <meta name="author" content="Gruppe D"> <!-- Author of Website -->
+
+    <!-- Visuals -->
+    <title>Olsker Cupcakes</title>
+
+    <!-- Icon -->
+    <link rel="icon" href="/images/icons/cupcake-1.png" type="image/x-icon"> <!-- That little logo thats next to the title in URL -->
+    <link rel="apple-touch-icon" href="/images/icons/cupcake-1.png">
+
+    <!-- CSS Load -->
+    <link rel="stylesheet" href="/css/style.css">
+
+    <!-- Scripts can also be loaded here with defer -->
+    <!-- Framwork load -->
+    <link rel="stylesheet" href="/css/guacamoleboy-wow.css">
+    <link rel="stylesheet" href="/css/guacamoleboy-framework.css">
+    <script src="/js/guacamoleboy-wow.js" defer></script>
+
+    <!-- Modules -->
+    <script type="module" src="/js/components/navbar.js" defer></script>
+
+</head>
+<body class="subpage">
+
+    <!-- Navbar -->
+    <div id="navbar-component"></div>
+    <!-- Navbar -->
+
+    <!-- Hero Section Start -->
+    <section class="hero">
+
+        <div class="guac-container">
+            <div class="guac-row">
+                
+                <div class="guac-col-auto subpage-cupcake-1">
+                    <img src="/images/page-content/index/cupcake-3.png" alt="Billede 1">
+                </div>
+                <div class="guac-col-auto subpage-cupcake-2">
+                    <img src="/images/page-content/index/cupcake-3.png" alt="Billede 2">
+                </div>
+
+            </div>
+        </div>
+
+        <!-- End Wave Hero Start -->
+        <div class="wave-1">
+            <img src="/images/waves/wave-top.svg">
+        </div>
+        <!-- End Wave Hero End -->
+
+    </section>
+    <!-- Hero Section End -->
+
+    <!-- Next Section Start -->
+    <section class="section-1">
+        <div class="guac-container">
+            <div class="guac-row">
+
+            </div>
+        </div>
+    </section>
+    <!-- Next Section End -->
+
+</body>
+</html>
