+++ Cupcake/src/main/resources/templates/login.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html lang="da">
+<head>
+
+    <!-- Meta -->
+    <meta charset="UTF-8"> <!-- ÆØÅ -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"> <!-- In case some nerd uses Internet Explorer -->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- Mobile Responsiveness -->
+    <meta name="description" content="Fuck vi hader livet jo... #swag"> <!-- SEO / Google -->
+    <meta name="author" content="Gruppe D"> <!-- Author of Website -->
+
+    <!-- Visuals -->
+    <title>Olsker Cupcakes</title>
+
+    <!-- Icon -->
+    <link rel="icon" href="/images/icons/cupcake-1.png" type="image/x-icon"> <!-- That little logo thats next to the title in URL -->
+    <link rel="apple-touch-icon" href="/images/icons/cupcake-1.png">
+
+    <!-- CSS Load -->
+    <link rel="stylesheet" href="/css/style.css">
+
+    <!-- Framwork load -->
+    <link rel="stylesheet" href="/css/guacamoleboy-wow.css">
+    <link rel="stylesheet" href="/css/guacamoleboy-framework.css">
+    <script src="/js/guacamoleboy-wow.js"></script>
+
+    <!-- Font Load -->
+    <link href="https://fonts.googleapis.com/css2?family=Itim&display=swap" rel="stylesheet">
+
+    <!-- Script Load -->
+    <script src="/js/bubble-spawn.js" defer></script>
+
+</head>
+<body> 
+
+    <a href="/" class="exit-btn" title="Tilbage til forsiden">&#10006;</a>
+
+    <section class="login-section-1">
+        <div class="login-container">
+            <h2>Log Ind</h2>
+
+            <form method="post" action="/login">
+                <input type="text" name="username" placeholder="Brugernavn" required>
+                <input type="password" name="password" placeholder="Adgangskode" required>
+
+                <div class="login-btn-group">
+                    <button type="submit" class="guac-btn guac-btn-form">
+                        Log Ind
+                    </button>
+                    <button type="button" class="guac-btn guac-btn-form" onclick="window.location.href='/register'">
+                        Opret Bruger
+                    </button>
+                </div>
+
+                <p th:if="${error}" th:text="${error}" style="color: red;"></p>
+            </form>
+        </div>
+    </section>
+
+</body>
+</html>
