+++ Cupcake/src/main/resources/templates/profile.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="da">
+<head>
+
+    <!-- Meta -->
+    <meta charset="UTF-8"> <!-- ÆØÅ -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"> <!-- In case some nerd uses Internet Explorer -->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- Mobile Responsiveness -->
+    <meta name="description" content="Fuck vi hader livet jo... #swag"> <!-- SEO / Google -->
+    <meta name="author" content="Gruppe D"> <!-- Author of Website -->
+
+    <!-- Visuals -->
+    <title>Olsker Cupcakes</title>
+
+    <!-- Icon -->
+    <link rel="icon" href="/images/icons/cupcake-1.png" type="image/x-icon"> <!-- That little logo thats next to the title in URL -->
+    <link rel="apple-touch-icon" href="/images/icons/cupcake-1.png">
+
+    <!-- CSS Load -->
+    <link rel="stylesheet" href="/css/style.css">
+
+    <!-- Scripts can also be loaded here with defer -->
+    <!-- Framwork load -->
+    <link rel="stylesheet" href="/css/guacamoleboy-wow.css">
+    <link rel="stylesheet" href="/css/guacamoleboy-framework.css">
+    <script src="/js/guacamoleboy-wow.js" defer></script>
+
+    <!-- Modules -->
+    <script type="module" src="/js/components/navbar.js" defer></script>
+
+</head>
+<body class="profile-initial">
+
+    <!-- Profile Menu Start -->
+    <section class="profile-menu-visuals">
+        <div class="guac-row">
+
+            <!-- Side Menu Start -->
+            <div class="guac-col-2 profile-nav">
+                <div class="profile-title">
+                    <h3>Hej,<br>$name!</h3>
+                </div>
+                <div class="profile-buttons">
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                    <a href="#" class="guac-btn guac-btn-profile">Option</a>
+                </div>
+
+                <!-- Exit Button Start -->
+                <a href="../templates/index.html" class="exit-btn-profile" title="Tilbage til forsiden">&#10006;</a>
+                <!-- Exit Button End -->
+
+            </div>
+            <!-- Side Menu End -->
+
+            <!-- Main Content Start -->
+            <div class="guac-col-auto profile-actual">
+                <h2>Ku' fandme godt lægge en lort lige nu</h2>
+            </div>
+            <!-- Main Content End -->
+
+        </div>
+    </section>
+    <!-- Profile Menu End -->
+
+</body>
+</html>
