+++ Cupcake/src/main/resources/templates/register.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="da">
+<head>
+
+    <!-- Meta -->
+    <meta charset="UTF-8"> <!-- ÆØÅ -->
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"> <!-- In case some nerd uses Internet Explorer -->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- Mobile Responsiveness -->
+    <meta name="description" content="Fuck vi hader livet jo... #swag"> <!-- SEO / Google -->
+    <meta name="author" content="Gruppe D"> <!-- Author of Website -->
+
+    <!-- Visuals -->
+    <title>Olsker Cupcakes</title>
+
+    <!-- Icon -->
+    <link rel="icon" href="/images/icons/cupcake-1.png" type="image/x-icon"> <!-- That little logo thats next to the title in URL -->
+    <link rel="apple-touch-icon" href="/images/icons/cupcake-1.png">
+
+    <!-- CSS Load -->
+    <link rel="stylesheet" href="/css/style.css">
+
+    <!-- Framwork load -->
+    <link rel="stylesheet" href="/css/guacamoleboy-wow.css">
+    <link rel="stylesheet" href="/css/guacamoleboy-framework.css">
+    <script src="/js/guacamoleboy-wow.js"></script>
+
+    <!-- Script Load -->
+    <script src="/js/bubble-spawn.js" defer></script>
+
+</head>
+<body>
+
+    <!-- Exit Button Start -->
+    <a href="/" class="exit-btn" title="Tilbage til forsiden">&#10006;</a>
+    <!-- Exit Button End -->
+
+    <!-- Register Section Start -->
+    <section class="login-section-1">
+        <div class="login-container">
+            <h2>Opret Bruger</h2>
+            <form id="registerForm" method="post" action ="/register">
+                <input type="text" name ="username" placeholder="Brugernavn" required>
+                <input type="password" name ="password" placeholder="Adgangskode" required>
+                <input type="email" name ="email" placeholder="Email" required>
+                <div class="login-btn-group">
+                    <button type="button" class="guac-btn guac-btn-form" onclick="window.location.href='/login'">
+                        Tilbage
+                    </button>
+                    <button type="submit" class="guac-btn guac-btn-form">
+                        Opret Bruger
+                    </button>
+                </div>
+                <p th:if="${error}" th:text="${error}" style="color: red;"></p>
+            </form>
+        </div>
+    </section>
+    <!-- Register Section End -->
+
+</body>
+</html>
